Removed dependency cache code.
