+++ src/main/java/org/craftercms/core/service/CacheService.java
@@ -87,22 +87,10 @@ public interface CacheService {
     void put(Context context, Object key, Object value) throws InvalidContextException, InternalCacheEngineException;
 
     /**
-     * Puts an item in the scope of the given context.
-     */
-    void put(Context context, Object key, Object value, List<Object> dependencyKeys) throws InvalidScopeException,
-        InternalCacheEngineException;
-
-    /**
      * Puts and item in the scope of the given context.
      */
     void put(Context context, Object key, Object value, CachingOptions cachingOptions, CacheLoader loader,
              Object... loaderParams) throws InvalidContextException, InternalCacheEngineException;
-
-    /**
-     * Puts and item in the scope of the given context.
-     */
-    void put(Context context, Object key, Object value, List<Object> dependencyKeys, CachingOptions cachingOptions,
-             CacheLoader loader, Object... loaderParams) throws InvalidContextException, InternalCacheEngineException;
 
     /**
      * Removes an item from the scope of the given context.
