+++ src/main/java/org/craftercms/core/service/impl/ContentStoreServiceImpl.java
@@ -254,11 +254,6 @@ public class ContentStoreServiceImpl extends AbstractCachedContentStoreService {
             if (item.getDescriptorDom() != null) {
                 item = doMerging(context, cachingOptions, item);
                 item = doProcessing(context, cachingOptions, item, processor);
-            } else {
-                // Since there was no processing, add the original key (from the store adapter item) as dependency key.
-                // The store
-                // service item key will be set later.
-                item.addDependencyKey(item.getKey());
             }
         }
 
@@ -303,7 +298,6 @@ public class ContentStoreServiceImpl extends AbstractCachedContentStoreService {
                                                                                              processor);
                 if (treeChildren != null) {
                     tree.setChildren(treeChildren.getActualList());
-                    tree.addDependencyKeys(treeChildren.getDependencyKeys());
                 }
             }
 
@@ -331,9 +325,6 @@ public class ContentStoreServiceImpl extends AbstractCachedContentStoreService {
         XmlFileParseException, XmlMergeException, ItemProcessingException, StoreException {
         List<Item> children = context.getStoreAdapter().findItems(context, cachingOptions, url, false);
         if (children != null) {
-            List<Object> dependencyKeys = new ArrayList<>();
-            dependencyKeys.add(((CachingAwareList<Item>)children).getKey());
-
             if (filter != null && filter.runBeforeProcessing()) {
                 if (logger.isDebugEnabled()) {
                     logger.debug("Running filter " + filter + " before processing for " + url + "...");
@@ -365,14 +356,7 @@ public class ContentStoreServiceImpl extends AbstractCachedContentStoreService {
 
             Collections.sort(processedChildren, CompareByItemNameComparator.instance);
 
-            for (Item child : processedChildren) {
-                dependencyKeys.add(child.getKey());
-            }
-
-            CachingAwareList<Item> finalChildren = new CachingAwareList<Item>(processedChildren);
-            finalChildren.setDependencyKeys(dependencyKeys);
-
-            return finalChildren;
+            return new CachingAwareList<>(processedChildren);
         } else {
             return null;
         }
@@ -433,8 +417,6 @@ public class ContentStoreServiceImpl extends AbstractCachedContentStoreService {
                 }
 
                 documentsToMerge.add(descriptorDom);
-
-                item.addDependencyKey(descriptorItem.getKey());
             } else if (!descriptorToMerge.isOptional()) {
                 throw new XmlMergeException("Descriptor file " + descriptorUrl + " not found and is marked as " +
                                             "required for merging");
