+++ src/main/java/org/craftercms/core/store/impl/AbstractFileBasedContentStoreAdapter.java
@@ -175,7 +175,6 @@ public abstract class AbstractFileBasedContentStoreAdapter extends AbstractCache
 
                     if (item != null) {
                         items.add(item);
-                        items.addDependencyKey(item.getKey());
                     }
                 }
             }
